Implement method

+++ src/main/java/application/minigame/memory/SecretCode.java
@@ -6,12 +6,12 @@ public interface SecretCode {
      * Generates a random Secret Code.
      * @return a random long.
      */
-    Long generateSecretCode();
+    int generateSecretCode();
 
     /**
      * Confront secretCode with user input.
      * @param secretCode
      * @return the number of errors.
      */
-    int checkCode(Long secretCode);
+    int checkCode(int secretCode);
 }
